+++ common/src/main/java/com/mlib/client/CustomParticle.java
@@ -2,14 +2,18 @@ package com.mlib.client;
 
 import com.mlib.annotation.Dist;
 import com.mlib.annotation.OnlyIn;
-import com.mlib.mixininterfaces.IMixinSingleQuadParticle;
+import com.mojang.blaze3d.vertex.VertexConsumer;
+import net.minecraft.client.Camera;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.particle.*;
 import net.minecraft.core.particles.SimpleParticleType;
+import net.minecraft.util.Mth;
+import net.minecraft.world.phys.Vec3;
 import org.joml.Quaternionf;
+import org.joml.Vector3f;
 
 @OnlyIn( Dist.CLIENT )
-public abstract class CustomParticle extends TextureSheetParticle implements IMixinSingleQuadParticle {
+public abstract class CustomParticle extends TextureSheetParticle {
 	public IFormula< Double > xdFormula = xd->xd * ( this.onGround ? 0.5 : 0.95 );
 	public IFormula< Double > ydFormula = yd->yd - ( this.onGround ? 0.0 : 0.0375 );
 	public IFormula< Double > zdFormula = zd->zd * ( this.onGround ? 0.5 : 0.95 );
@@ -38,6 +42,66 @@ public abstract class CustomParticle extends TextureSheetParticle implements IMi
 	}
 
 	@Override
+	public void render( VertexConsumer consumer, Camera camera, float partialTick ) {
+		// it overwrites default behavior just to modify y and rotation
+		// but it is not using mixin to modify them because other performance mods
+		// like sodium also overwrites the code which makes the game crash
+		Vec3 vec3 = camera.getPosition();
+		float x = ( float )( Mth.lerp( partialTick, this.xo, this.x ) - vec3.x() );
+		float y = this.getY( ( float )( Mth.lerp( partialTick, this.yo, this.y ) - vec3.y() ) );
+		float z = ( float )( Mth.lerp( partialTick, this.zo, this.z ) - vec3.z() );
+		Quaternionf quaternion;
+		if( this.roll == 0.0F ) {
+			quaternion = camera.rotation();
+		} else {
+			quaternion = new Quaternionf( camera.rotation() );
+			quaternion.rotateZ( Mth.lerp( partialTick, this.oRoll, this.roll ) );
+		}
+		quaternion = this.getQuaternion( quaternion );
+
+		Vector3f[] avector3f = new Vector3f[]{
+			new Vector3f( -1.0f, -1.0f, 0.0f ),
+			new Vector3f( -1.0f, 1.0f, 0.0f ),
+			new Vector3f( 1.0f, 1.0f, 0.0f ),
+			new Vector3f( 1.0f, -1.0f, 0.0f )
+		};
+		float size = this.getQuadSize( partialTick );
+
+		for( int i = 0; i < 4; ++i ) {
+			Vector3f vector3f = avector3f[ i ];
+			vector3f.rotate( quaternion );
+			vector3f.mul( size );
+			vector3f.add( x, y, z );
+		}
+
+		float f6 = this.getU0();
+		float f7 = this.getU1();
+		float f4 = this.getV0();
+		float f5 = this.getV1();
+		int light = this.getLightColor( partialTick );
+		consumer.vertex( avector3f[ 0 ].x(), avector3f[ 0 ].y(), avector3f[ 0 ].z() )
+			.uv( f7, f5 )
+			.color( this.rCol, this.gCol, this.bCol, this.alpha )
+			.uv2( light )
+			.endVertex();
+		consumer.vertex( avector3f[ 1 ].x(), avector3f[ 1 ].y(), avector3f[ 1 ].z() )
+			.uv( f7, f4 )
+			.color( this.rCol, this.gCol, this.bCol, this.alpha )
+			.uv2( light )
+			.endVertex();
+		consumer.vertex( avector3f[ 2 ].x(), avector3f[ 2 ].y(), avector3f[ 2 ].z() )
+			.uv( f6, f4 )
+			.color( this.rCol, this.gCol, this.bCol, this.alpha )
+			.uv2( light )
+			.endVertex();
+		consumer.vertex( avector3f[ 3 ].x(), avector3f[ 3 ].y(), avector3f[ 3 ].z() )
+			.uv( f6, f5 )
+			.color( this.rCol, this.gCol, this.bCol, this.alpha )
+			.uv2( light )
+			.endVertex();
+	}
+
+	@Override
 	public ParticleRenderType getRenderType() {
 		return ParticleRenderType.PARTICLE_SHEET_OPAQUE;
 	}
@@ -47,12 +111,10 @@ public abstract class CustomParticle extends TextureSheetParticle implements IMi
 		return this.quadSize * this.scaleFormula.apply( ( ( float )this.age + scaleFactor ) / ( float )this.lifetime );
 	}
 
-	@Override
 	public float getY( float y ) {
 		return y;
 	}
 
-	@Override
 	public Quaternionf getQuaternion( Quaternionf quaternion ) {
 		return quaternion;
 	}