+++ common/src/main/java/com/mlib/mixininterfaces/IMixinSingleQuadParticle.java
@@ -1,9 +1,0 @@
-package com.mlib.mixininterfaces;
-
-import org.joml.Quaternionf;
-
-public interface IMixinSingleQuadParticle {
-	float getY( float y );
-
-	Quaternionf getQuaternion( Quaternionf quaternion );
-}